Add logo to top of README
Not necessary but it's fun :D

This embeds the logo on spigotmc.org, but I'm not sure what happens if the logo there changes. This might break and have to get updated too, or it might automatically update if the file gets replaced instead of deleted and a new one added.